Create page-specific styles
Made a discovery in globals.css \\ realized I haven't enclosed almost
all of the selectors, but it still works. Define selectors inside a
directive to specify which "bucket" a set of custom styles belong to, as
it leads to more control and optimization in final build.

Also discovered the @tailwind directive resets global styles, perhaps it
is best to define non-enclosed selectors when performing changes on top
of global styles. I'd reckon it would be unideal long-term.