[ERR-1] Revert display prop to default of one svg
Wrong theory; inline-block does not allow for place-self-center to work
as place-self is designed for layout modules like flex and grid.

The svgs are block by default, so mx-auto is the alt solution. Setting
to inline/inline-block causes the centering to be affected by
text-center which is why it worked.