Revert component import in page
Module calls seem to be interefering with styles, and importing the
carousel directly to page sends a compile error. I removed the styling
imports for the components.

Figured that when I'm designing a component to be rendered on main page,
I should call it on the main to view realtime changes rather than
compiling the component with imported styles.